src/requirements.txt
braintree==3.31.0
Django==1.10.3
django-allauth==0.29.0
django-crispy-forms==1.6.1
django-filter==1.0.0
django-markdown-deux==1.0.5
django-pagedown==0.1.1
markdown2==2.3.1
oauthlib==2.0.0
Pillow==3.4.2
psycopg2==2.6.2
python-openid==2.2.5
requests==2.12.1
requests-oauthlib==0.7.0
